+++ src/repo/reservation.ts
@@ -178,6 +178,19 @@ export class MongoRepository {
                         }
                     });
                 }
+
+                const additionalTicketTextRegex = params.additionalTicketText?.$regex;
+                if (typeof additionalTicketTextRegex === 'string') {
+                    const additionalTicketTextOptions = params.additionalTicketText?.$options;
+
+                    andConditions.push({
+                        additionalTicketText: {
+                            $exists: true,
+                            $regex: new RegExp(additionalTicketTextRegex),
+                            ...(typeof additionalTicketTextOptions === 'string') ? { $options: additionalTicketTextOptions } : undefined
+                        }
+                    });
+                }
             }
         }
 
@@ -546,18 +559,16 @@ export class MongoRepository {
 
                 // tslint:disable-next-line:no-single-line-block-comment
                 /* istanbul ignore else */
-                if ((<any>params.reservedTicket.ticketedSeat).seatingType !== undefined
-                    && (<any>params.reservedTicket.ticketedSeat).seatingType !== null) {
-                    if (Array.isArray((<any>params.reservedTicket.ticketedSeat).seatingType.$in)) {
-                        andConditions.push(
-                            {
-                                'reservedTicket.ticketedSeat.seatingType': {
-                                    $exists: true,
-                                    $in: (<any>params.reservedTicket.ticketedSeat).seatingType.$in
-                                }
+                const ticketedSeatSeatingTypeIn = params.reservedTicket.ticketedSeat?.seatingType?.$in;
+                if (Array.isArray(ticketedSeatSeatingTypeIn)) {
+                    andConditions.push(
+                        {
+                            'reservedTicket.ticketedSeat.seatingType': {
+                                $exists: true,
+                                $in: ticketedSeatSeatingTypeIn
                             }
-                        );
-                    }
+                        }
+                    );
                 }
             }
         }
@@ -578,24 +589,50 @@ export class MongoRepository {
 
             // tslint:disable-next-line:no-single-line-block-comment
             /* istanbul ignore else */
-            if (params.underName.email !== undefined) {
+            if (typeof params.underName.email === 'string') {
                 andConditions.push({
                     'underName.email': {
                         $exists: true,
                         $regex: new RegExp(params.underName.email)
                     }
                 });
+            } else {
+                const emailRegex = params.underName.email?.$regex;
+                if (typeof emailRegex === 'string') {
+                    const emailOptions = params.underName.email?.$options;
+
+                    andConditions.push({
+                        'underName.email': {
+                            $exists: true,
+                            $regex: new RegExp(emailRegex),
+                            ...(typeof emailOptions === 'string') ? { $options: emailOptions } : undefined
+                        }
+                    });
+                }
             }
 
             // tslint:disable-next-line:no-single-line-block-comment
             /* istanbul ignore else */
-            if (params.underName.name !== undefined) {
+            if (typeof params.underName.name === 'string') {
                 andConditions.push({
                     'underName.name': {
                         $exists: true,
                         $regex: new RegExp(params.underName.name)
                     }
                 });
+            } else {
+                const underNameNameRegex = params.underName.name?.$regex;
+                if (typeof underNameNameRegex === 'string') {
+                    const underNameNameOptions = params.underName.name?.$options;
+
+                    andConditions.push({
+                        'underName.name': {
+                            $exists: true,
+                            $regex: new RegExp(underNameNameRegex),
+                            ...(typeof underNameNameOptions === 'string') ? { $options: underNameNameOptions } : undefined
+                        }
+                    });
+                }
             }
 
             // tslint:disable-next-line:no-single-line-block-comment
@@ -611,24 +648,50 @@ export class MongoRepository {
 
             // tslint:disable-next-line:no-single-line-block-comment
             /* istanbul ignore else */
-            if (params.underName.givenName !== undefined) {
+            if (typeof params.underName.givenName === 'string') {
                 andConditions.push({
                     'underName.givenName': {
                         $exists: true,
                         $regex: new RegExp(params.underName.givenName)
                     }
                 });
+            } else {
+                const givenNameRegex = params.underName.givenName?.$regex;
+                if (typeof givenNameRegex === 'string') {
+                    const givenNameOptions = params.underName.givenName?.$options;
+
+                    andConditions.push({
+                        'underName.givenName': {
+                            $exists: true,
+                            $regex: new RegExp(givenNameRegex),
+                            ...(typeof givenNameOptions === 'string') ? { $options: givenNameOptions } : undefined
+                        }
+                    });
+                }
             }
 
             // tslint:disable-next-line:no-single-line-block-comment
             /* istanbul ignore else */
-            if (params.underName.familyName !== undefined) {
+            if (typeof params.underName.familyName === 'string') {
                 andConditions.push({
                     'underName.familyName': {
                         $exists: true,
                         $regex: new RegExp(params.underName.familyName)
                     }
                 });
+            } else {
+                const familyNameRegex = params.underName.familyName?.$regex;
+                if (typeof familyNameRegex === 'string') {
+                    const familyNameOptions = params.underName.familyName?.$options;
+
+                    andConditions.push({
+                        'underName.familyName': {
+                            $exists: true,
+                            $regex: new RegExp(familyNameRegex),
+                            ...(typeof familyNameOptions === 'string') ? { $options: familyNameOptions } : undefined
+                        }
+                    });
+                }
             }
 
             // tslint:disable-next-line:no-single-line-block-comment
@@ -742,6 +805,20 @@ export class MongoRepository {
         return andConditions;
     }
 
+    public async distinct<T extends factory.reservationType>(
+        field: string,
+        params: factory.reservation.ISearchConditions<T>
+    ): Promise<any[]> {
+        const conditions = MongoRepository.CREATE_MONGO_CONDITIONS(params);
+
+        return this.reservationModel.distinct(
+            field,
+            (conditions.length > 0) ? { $and: conditions } : {}
+        )
+            .setOptions({ maxTimeMS: 10000 })
+            .exec();
+    }
+
     /**
      * 汎用予約カウント
      */
@@ -756,19 +833,22 @@ export class MongoRepository {
     }
 
     /**
-     * 汎用予約検索
+     * 予約検索
      */
     public async search<T extends factory.reservationType>(
-        params: factory.reservation.ISearchConditions<T>
+        params: factory.reservation.ISearchConditions<T>,
+        projection?: any
     ): Promise<factory.reservation.IReservation<factory.reservationType.EventReservation>[]> {
         const conditions = MongoRepository.CREATE_MONGO_CONDITIONS(params);
         const query = this.reservationModel.find(
             (conditions.length > 0) ? { $and: conditions } : {},
-            {
-                __v: 0,
-                createdAt: 0,
-                updatedAt: 0
-            }
+            (projection !== undefined && projection !== null)
+                ? projection
+                : {
+                    __v: 0,
+                    createdAt: 0,
+                    updatedAt: 0
+                }
         );
 
         // tslint:disable-next-line:no-single-line-block-comment
@@ -837,13 +917,23 @@ export class MongoRepository {
             previousReservationStatus?: factory.reservationStatusType;
         }
     ): Promise<factory.reservation.IReservation<T>> {
+        const update = {
+            ...params,
+            reservationStatus: factory.reservationStatusType.ReservationConfirmed,
+            modifiedTime: new Date()
+        };
+
+        // 以下値は更新しない(他処理で更新される可能性あり)
+        if (typeof update.checkedIn === 'boolean') {
+            delete update.checkedIn;
+        }
+        if (typeof update.attended === 'boolean') {
+            delete update.attended;
+        }
+
         const doc = await this.reservationModel.findByIdAndUpdate(
             String(params.id),
-            {
-                ...<any>params,
-                reservationStatus: factory.reservationStatusType.ReservationConfirmed,
-                modifiedTime: new Date()
-            },
+            <any>update,
             {
                 new: true
             }