+++ src/service/event.ts
@@ -3,7 +3,6 @@
  */
 import * as COA from '@motionpicture/coa-service';
 import * as createDebug from 'debug';
-// import { PhoneNumberFormat, PhoneNumberUtil } from 'google-libphonenumber';
 // @ts-ignore
 import * as difference from 'lodash.difference';
 // import { google } from 'googleapis';