+++ CHANGELOG.md
@@ -17,9 +17,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - アクション検索条件拡張
 - イベントに入場ゲート集計属性を追加
 - ID指定されたイベントについてはEventScheduledでなくても予約集計するように調整
-- 予約にuseActionExists属性を追加
 - 予約にreservedTicket.dateUsed属性を実験的に追加
 - 予約検索に$projectionを指定できるように調整
+- update packages
+- update typescript
 
 ### Deprecated